indeed_bot_pending_scripts/requirements_to_freeze.txt
google-api-python-client
oauth2client
black
selenium
fake_useragent
anticaptchaofficial
gspread
bs4
lxml
pyinstaller
python-dotenv
sqlalchemy
pycountry
selenium-wire